+++ src/main/java/com/example/junioroasisbackend/services/posts/PostServiceImpl.java
@@ -5,18 +5,16 @@ import com.example.junioroasisbackend.dtos.responses.CommentResponseDTO;
 import com.example.junioroasisbackend.dtos.responses.PostResponseDTO;
 import com.example.junioroasisbackend.entities.User;
 import com.example.junioroasisbackend.repositories.PostRepository;
-import com.example.junioroasisbackend.repositories.UserRepository;
 import com.example.junioroasisbackend.entities.Post;
 import com.example.junioroasisbackend.services.comments.CommentService;
+import com.example.junioroasisbackend.services.user.UserService;
 import com.example.junioroasisbackend.utils.Convertor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.*;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
-import java.util.Optional;
+import java.util.Objects;
 
 
 @Service
@@ -24,7 +22,7 @@ import java.util.Optional;
 public class PostServiceImpl implements PostService {
 
 
-    private final UserRepository userRepository;
+    private final UserService userService;
 
     private final PostRepository postRepository;
 
@@ -32,15 +30,15 @@ public class PostServiceImpl implements PostService {
     private final CommentService commentService;
 
 
+
     @Override
     public PostResponseDTO addPost(PostRequestDTO postDTO) {
-        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication()
-                .getPrincipal();
-        Optional<User> optionalUser = userRepository.findFirstByEmail(userDetails.getUsername());
-        return optionalUser.map(user -> PostResponseDTO.mapToDto(postRepository
-                        .save(this.assignPostRequestedToPost(postDTO, user))))
-                .orElse(null);
+        User user = userService.getCurrentUser();
+       return  PostResponseDTO.mapToDto(postRepository
+                .save(this.assignPostRequestedToPost(postDTO, user)));
     }
+
+
 
     @Override
     public Page<PostResponseDTO> getAllPosts(Integer pageNumber, Integer perPage, String sortBy) {
@@ -59,8 +57,17 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public void deletePostById(Long postId) {
-        postRepository.deleteById(postId);
+    public boolean deletePostById(Long postId) throws Exception {
+
+        Post post  = this.getPostById(postId);
+        User user = this.userService.getCurrentUser();
+        //Dans cette ligne, Objects.equals() est utilisé pour comparer les ID de l'utilisateur actuel (user.getId()) et de l'utilisateur associé au post (post.getUser().getId()).
+        //Objects.equals() est utilisé pour comparer les ID des utilisateurs de manière sûre, prenant en compte les cas où l'un ou l'autre ID pourrait être nul sans provoquer une exception NullPointerException. Cela permet de vérifier l'égalité des ID sans risque d'erreur due à des références nulles.
+        if (Objects.equals(user.getId(), post.getUser().getId())){
+            postRepository.deleteById(postId);
+            return true ;
+        }
+        return false ;
     }
 
     @Override
@@ -86,9 +93,16 @@ public class PostServiceImpl implements PostService {
 
         Post post = this.getPostById(postId);
         String[] attrs = {"id", "body", "createdAt"};
-        Page<CommentResponseDTO> pageObj = commentService.getAllCommentsByPost(postId, pageNumber, perPage, Arrays.asList(attrs).contains(sortBy) ? sortBy : "createdAt");
+        Page<CommentResponseDTO> pageObj = commentService.getAllCommentsByPost(
+                postId,
+                pageNumber,
+                perPage,
+                Arrays.asList(attrs) //Arrays.asList(attrs) prend le tableau attrs (qui contient des chaînes de caractères) et le convertit en une liste (List<String>).
+                        .contains(sortBy) ? sortBy : "createdAt"); // Vérifie si sortBy est un attribut valide, sinon utilise "createdAt"
         PostResponseDTO postResponseDTO = PostResponseDTO.mapToDto(post);
         postResponseDTO.setComments(pageObj);
         return postResponseDTO;
     }
+
+
 }
